Big refactoring
Removing AlamofireRSSParser and using XMLParsing instead.
AlamofireRSSParser gives big object that contains almost every element of the RSS. Since there is no need for all of the data in this app, XMLParsing is used instead. That gives us control over which parameters will be read. Since items in RSS don't include images, in list of items for feed, feed image is displayed.